Move test utils to seperate files
